+++ Source/Plugin.BLE/Shared/ServiceBase.cs
@@ -7,23 +7,47 @@ using Plugin.BLE.Abstractions.Contracts;
 
 namespace Plugin.BLE.Abstractions
 {
+    /// <summary>
+    /// Base class for platform-specific <c>Service</c> classes.
+    /// </summary>
     public abstract class ServiceBase<TNativeService> : IService
     {
         private readonly List<ICharacteristic> _characteristics = new List<ICharacteristic>();
 
+        /// <summary>
+        /// Name of the service.
+        /// </summary>
         public string Name => KnownServices.Lookup(Id).Name;
+        /// <summary>
+        /// Id of the Service.
+        /// </summary>
         public abstract Guid Id { get; }
+        /// <summary>
+        /// Indicates whether the type of service is primary or secondary.
+        /// </summary>
         public abstract bool IsPrimary { get; }
+        /// <summary>
+        /// The parent device.
+        /// </summary>
         public IDevice Device { get; }
+        /// <summary>
+        /// The native service.
+        /// </summary>
         protected TNativeService NativeService { get; }
 
+        /// <summary>
+        /// ServiceBase constructor.
+        /// </summary>
         protected ServiceBase(IDevice device, TNativeService nativeService)
         {
             Device = device;
             NativeService = nativeService;
         }
 
-        public async Task<IReadOnlyList<ICharacteristic>> GetCharacteristicsAsync(CancellationToken cancellationToken = default(CancellationToken))
+        /// <summary>
+        /// Gets the characteristics of the service.
+        /// </summary>
+        public async Task<IReadOnlyList<ICharacteristic>> GetCharacteristicsAsync()
         {
             if (!_characteristics.Any())
             {
@@ -34,14 +58,24 @@ namespace Plugin.BLE.Abstractions
             return _characteristics.ToList();
         }
 
-        public async Task<ICharacteristic> GetCharacteristicAsync(Guid id, CancellationToken cancellationToken = default(CancellationToken))
+        /// <summary>
+        /// Gets the first characteristic with the Id <paramref name="id"/>. 
+        /// </summary>
+        /// <param name="id">The id of the searched characteristic.</param>
+        public async Task<ICharacteristic> GetCharacteristicAsync(Guid id)
         {
             var characteristics = await GetCharacteristicsAsync(cancellationToken);
             return characteristics.FirstOrDefault(c => c.Id == id);
         }
 
-        protected abstract Task<IList<ICharacteristic>> GetCharacteristicsNativeAsync(CancellationToken cancellationToken = default(CancellationToken));
+        /// <summary>
+        /// Native implementation of <c>GetCharacteristicsAsync</c>.
+        /// </summary>
+        protected abstract Task<IList<ICharacteristic>> GetCharacteristicsNativeAsync();
 
+        /// <summary>
+        /// Dispose the service.
+        /// </summary>
         public virtual void Dispose()
         {