+++ Source/Plugin.BLE/Shared/Trace.cs
@@ -2,12 +2,19 @@ using System;
 
 namespace Plugin.BLE.Abstractions
 {
+    /// <summary>
+    /// Generic tracing class.
+    /// </summary>
     public static class Trace
     {
-        public static Action<string, object[]> TraceDebugImplementation { get; set; }
+        /// <summary>
+        /// The actual tracing implementation.
+        /// </summary>
+        public static Action<string, object[]> TraceImplementation { get; set; }
 
-        public static Action<string, object[]> TraceInfoImplementation { get; set; }
-
+        /// <summary>
+        /// Print a message via the tracing implementation.
+        /// </summary>
         public static void Message(string format, params object[] args)
         {
             try